remove metadata from valor_lite.text_generation
